comin numero 10

+++ api/views.py
@@ -1,60 +1,74 @@
-from .serializer import LoginSerializer, ProfesionalServicioRelacionSerializer, ProfessionalImageSerializer, UserSerializer,ServicioSerializer,ProfesionalServicioSerializer,ProfesionalSerializer,UserProfesionalSerializer,CiudadSerializer,ProvinciaSerializer,PaisSerializer
+from .serializers import LoginSerializer, ProfesionalServicioRelacionSerializer, ProfessionalImageSerializer, UserSerializer,ServicioSerializer,ProfesionalServicioSerializer,ProfesionalSerializer,UserProfesionalSerializer,CiudadSerializer,ProvinciaSerializer,PaisSerializer
 from rest_framework.response import Response
 from rest_framework import status,viewsets
 from django.contrib.auth import authenticate, login
 from .models import ProfesionalServicio, ProfessionalImage, User, Servicio,Profesional,Ciudad,Provincia,Pais
 from rest_framework.views import APIView
-from django.middleware.csrf import get_token
-from django.shortcuts import render,redirect
-from rest_framework import generics
-from django.urls import reverse_lazy
-from django.utils.decorators import method_decorator
-from django.views.decorators.cache import never_cache
-from django.views.decorators.csrf import csrf_protect
 from django.views.generic.edit import FormView
 from django.contrib.auth import login,logout,authenticate
-from django.http import HttpResponseRedirect
-from django.contrib.auth.forms import AuthenticationForm
 from rest_framework.authtoken.models import Token
-from rest_framework.permissions import IsAuthenticated
-from rest_framework.authentication import TokenAuthentication
 from rest_framework.views import APIView
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.permissions import AllowAny  # Importa AllowAny
+import json
 #login
 # Create your views here.
-
-#Login
+#api\views.py
+#Login 
 class LoginView(APIView):
     permission_classes = [AllowAny]
-
     def post(self, request):
-        serializer = LoginSerializer(data=request.data)
-        if serializer.is_valid():
-            user = serializer.validated_data
+        # Verificar si el usuario ya está autenticado
+        if request.user.is_authenticated:
+            user = request.user
+            token, created = Token.objects.get_or_create(user=user)
+            user_data = UserSerializer(user).data
+            user_data['token'] = token.key
+
+            if created:
+                return Response({"user": user_data, "detail": "Token creado exitosamente."}, status=status.HTTP_201_CREATED)
+            else:
+                return Response({"user": user_data, "detail": "Usuario ya estaba logueado."}, status=status.HTTP_200_OK)
+
+        # Verificar las credenciales del usuario
+        username = request.data.get('username')
+        password = request.data.get('password')
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
             login(request, user)
-            
-            # Obtener el token del usuario
-            try:
-                token = Token.objects.get(user=user)
-                user_data = UserSerializer(user).data
-                user_data['token'] = token.key
-                return Response({"user": user_data}, status=status.HTTP_200_OK)
-            except Token.DoesNotExist:
-                return Response({"detail": "El token del usuario no se encontró."}, status=status.HTTP_400_BAD_REQUEST)
+
+            # Obtener o crear el token del usuario
+            token, created = Token.objects.get_or_create(user=user)
+
+            # Serializar los datos del usuario y crear la respuesta
+            user_data = UserSerializer(user).data
+            user_data['token'] = token.key
+
+            if created:
+                return Response({"user": user_data, "detail": "Token creado exitosamente."}, status=status.HTTP_201_CREATED)
+            else:
+                return Response({"user": user_data, "detail": "Usuario ya estaba logueado."}, status=status.HTTP_200_OK)
         else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"detail": "Credenciales inválidas."}, status=status.HTTP_400_BAD_REQUEST)
     
 class Logout(APIView):
-    def get(self,request, format = None):
-        request.user.auth_token.delete()
-        logout(request)
-        return Response(status = status.HTTP_200_OK)
+    def get(self, request, format=None):
+        # Verificar si el usuario está autenticado
+        if request.user.is_authenticated:
+            # Eliminar el token de autenticación del usuario
+            request.user.auth_token.delete()
+            # Desautenticar al usuario
+            logout(request)
+            return Response({"detail": "Usuario deslogueado exitosamente."}, status=status.HTTP_200_OK)
+        else:
+            return Response({"detail": "No hay usuario autenticado."}, status=status.HTTP_400_BAD_REQUEST)
        
 class UserViewSet(viewsets.ModelViewSet):
+    permission_classes = [AllowAny]
     queryset = User.objects.all()
     serializer_class = UserSerializer
+    
 
 class ServicioViewSet(viewsets.ModelViewSet):  
     queryset = Servicio.objects.all()
@@ -69,6 +83,7 @@ class ProfesionalServicioViewSet(viewsets.ModelViewSet):
     serializer_class = ProfesionalServicioSerializer
 
 class ProfesionalViewSet(viewsets.ModelViewSet):
+    permission_classes = [AllowAny] #Quitar luego xd
     queryset = Profesional.objects.all()
     serializer_class = ProfesionalSerializer
